requirements.txt
# Core Dependencies
python>=3.10
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
deepspeed>=0.12.0

# Vector Store & Retrieval
chromadb>=0.4.22
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4

# Langfuse Tracking
langfuse>=2.0.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.1

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Fine-tuning
peft>=0.7.0  # For LoRA

# Development
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
